If this is what a dll is, then I'll take 10 plox.
pokin' brains

+++ src/Time.h
@@ -11,10 +11,10 @@ namespace MyTimer
 		__declspec(dllexport) Time(clock_t clock);
 		__declspec(dllexport) ~Time() {};
 
-		__declspec(dllexport) inline float Hours() { return ((float)(mClock / CLOCKS_PER_SEC)) / 60.0f / 60.0f; };
-		__declspec(dllexport) inline float Minutes() { return ((float)(mClock / CLOCKS_PER_SEC)) / 60.0f; };
-		__declspec(dllexport) inline float Seconds() { return ((float)(mClock / CLOCKS_PER_SEC)); };
-		__declspec(dllexport) inline float Milliseconds() { return ((float)(mClock / CLOCKS_PER_SEC)) * 1000.0f; };
+		__declspec(dllexport) inline float Hours() { return ((float)((float)mClock / (float)CLOCKS_PER_SEC)) / 60.0f / 60.0f; };
+		__declspec(dllexport) inline float Minutes() { return ((float)((float)mClock / (float)CLOCKS_PER_SEC)) / 60.0f; };
+		__declspec(dllexport) inline float Seconds() { return ((float)((float)mClock / (float)CLOCKS_PER_SEC)); };
+		__declspec(dllexport) inline float Milliseconds() { return ((float)((float)mClock / (float)CLOCKS_PER_SEC)) * 1000.0f; };
 	private:
 		clock_t mClock;
 	};
